requirements.txt
PyMuPDF
openpyxl
python-docx